refactor: moved some logic to temp hooks for better readability
